requirements.txt
Django>=2.2.2, <2.2.3
djangorestframework>=3.10.1,<3.10.2
psycopg2>=2.7.5,<2.8.0
PyYAML>=5.1.1,<5.1.2
uritemplate>=3.0.0
flake8>=3.6.0,<3.7.0
drf-yasg
packaging
